moved all token calls to the common code section
