+++ shared/exports.js
@@ -3,5 +3,5 @@ import { tokenRequest } from './token'
 
 export const common = {
   config: config,
-  tokenRequest: tokenRequest,
+  tokens: tokenRequest,
 }
